oplog_manager_py: Handle command entries from oplog for which the namepsace set must also include the respective $cmd namespace.
doc_managers/elastic_doc_manager.py: Add handle_command to handle database commands
and take respective action against the elasticsearch indices.  As of now,
only "drop" is handled.  More database commands needs to be handled.

Note: Upstream has a more wider implementation of this and it is still in beta of 2.0,
we needed a simplified implementation of this feature.