Use replica set client instead of standalone one
MongoClient assumes that the connected mongo instance is a standalone instance, which doesn't matter for the initial connection to the mongos/discovery instance, but is problematic for the replica set connection.

When specifying a read preference, it doesn't apply to the replica set connection, since they're standalone connections. If we were connected via mongos, that'd be no problem, since mongos would handle this, but with the standalone connection it just doesn't.

By using MongoReplicaSetClient, we ensure that the pymongo properly throws on read preference errors and also properly switches between instances.