Fix issue with nested properties
There were issues with updating/inserting documents, where nested
properties are not available in the index because they got filtered out
by the oplog manager.

e.g. if you specify to limit to the field "report.tags" of a given
collection, updating works fine when using $set in mongodb. When
inserting a complete new document to the collection the
pop_exclude_fields gets a full document (without dot-notation like with
$set), and behaves incorrect in terms of popping excluded fields (it
looks for "report.tags", but only "report" is in the document which gets
popped.