Do not use retry_until_ok in init_cursor
Instead wrap the call to init_cursor with retry_until_ok and handle failures
in the run method. This fixes a bug where an unwrapped call to
get_oldest_oplog_timestamp could cause the OplogThread to exit. It also makes
init_cursor easier to reason about and easier to debug.