Fixed ssl_cert_reqs, especially when 'ignored'
The config file setting "ssl.sslCertificatePolicy" was not being used to
set the corresponding option - the only value ever used was the value
from the command-line.

And even when the setting was passed in as on the command-line, it was
not being forwarded to the MongoClient if it was 'ignored,' since
'ignored' is mapped to 0 (boolean false).